fix: pr fix #2; refactor js utilities
