+++ app/javascript/packs/application.js
@@ -7,8 +7,7 @@ import Rails from "@rails/ujs"
 import Turbolinks from "turbolinks"
 import * as ActiveStorage from "@rails/activestorage"
 import "channels"
-import "utilities/sorting"
-import "utilities/input_check"
+import "utilities/manipulator"
 
 Rails.start()
 Turbolinks.start()
